requirements.txt
paho-mqtt
smbus2==0.4.3
